requirements.txt
langchain-groq>=0.1.0
langgraph>=0.0.20
streamlit>=1.28.0
pydantic>=2.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
langchain-core>=0.1.0
